api/requirements.txt
asgiref==3.8.1
Django==5.0.3
django-simple-history==3.5.0
django-softdelete==0.10.5
djangorestframework==3.15.1
mysql==0.0.3
mysqlclient==2.2.4
pillow==10.3.0
six==1.16.0
sqlparse==0.4.4
